Project_Data-Engineering/Task_1_1/requirements.txt
colorama==0.4.6
greenlet==2.0.2
numpy==1.25.1
pandas==2.0.3
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
